+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [Unreleased]
 ### Added
 - PASSBOLT-2694: Implement the webinstaller feature
+- PASSBOLT-2963: Persist the license in a file